requirements.txt
Flask
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
mysqlclient
python-dotenv
Werkzeug
PyJWT